+++ test/udp/server.js
@@ -4,7 +4,7 @@ var gn = require('../../src/gracenode');
 
 gn.config({
 	http: {
-		host: 'localhost',
+		host: '0.0.0.0',
 		port: 7979
 	},
 	udp: {
@@ -18,7 +18,7 @@ gn.config({
 });
 
 gn.http.post('/auth', handleAuth);
-gn.session.sessionDuration(1000);
+gn.session.sessionDuration(100000);
 gn.session.useUDPSession();
 gn.udp.hook(1, testHook1);
 gn.udp.command(1, 'testCommand1', testCommand1);